+++ NEWS.md
@@ -1,3 +1,4 @@
 # graph3d 1.0.0
 
-- s3d_addgrids(): Add grids to a scatterplot3d
+- s3d_addgrids(): Adds grids to a scatterplot3d
+- dist3d(): Draws a 3D histogram distribution of two variables